backend/requirements.txt
flask
azure-ai-formrecognizer
dotenv
azure-storage-blob
aiohttp
openai
twilio